feat: Finish rest api about order entity

+++ src/main/java/demo/bananeexport/controller/OrderController.java
@@ -4,11 +4,13 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import org.aspectj.weaver.ast.Or;
+import demo.bananeexport.exceptions.ApiCustomException;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,12 +18,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import demo.bananeexport.model.Order;
 import demo.bananeexport.repository.OrderRepository;
 
+import static org.springframework.data.domain.ExampleMatcher.GenericPropertyMatchers.exact;
 
 
 @RestController
@@ -35,9 +37,7 @@ class OrderController{
     public ResponseEntity<List<Order>> getAllOrders() {
         try {
             List<Order> orders = new ArrayList<Order>();
-
             orderRepository.findAll().forEach(orders::add);
-
             /*
             Si on veut gerer le HTTP CODE 204
             if (orders.isEmpty()) {
@@ -53,7 +53,6 @@ class OrderController{
     @GetMapping("/orders/{id}")
     public ResponseEntity<Order> getOrderById(@PathVariable("id") long id) {
         Optional<Order> orderData = orderRepository.findById(id);
-
         if (orderData.isPresent()) {
             return new ResponseEntity<>(orderData.get(), HttpStatus.OK);
         } else {
@@ -62,42 +61,58 @@ class OrderController{
     }
 
     @PostMapping("/orders")
-    public ResponseEntity<Order> createOrder(@RequestBody Order order) {
-        try {
-            Order _order = orderRepository
-                    .save(new Order(
-                            order.getName(),
-                            order.getAddress(),
-                            order.getPostal_code(),
-                            order.getCity(),
-                            order.getCountry()
-                        ));
-            return new ResponseEntity<>(_order, HttpStatus.CREATED);
-        } catch (Exception e) {
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+    public ResponseEntity<Order> createOrder(@Valid @RequestBody Order order) throws ApiCustomException{
+        ExampleMatcher modelMatcher = ExampleMatcher.matching()
+                .withIgnorePaths("id")
+                .withMatcher("name", exact())
+                .withMatcher("adress",exact())
+                .withMatcher("postal_code",exact())
+                .withMatcher("city",exact())
+                .withMatcher("country",exact())
+        ;
+        Example<Order> example = Example.of(order, modelMatcher);
+        boolean exists = orderRepository.exists(example);
+        if(exists){
+            throw  new ApiCustomException("User exist");
         }
+        System.out.println(order);
+        Order _order = orderRepository.save(order);
+        return new ResponseEntity<>(_order, HttpStatus.CREATED);
     }
 
     @PutMapping("/orders/{id}")
-    public ResponseEntity<Order> updateOrder(@PathVariable("id") long id, @RequestBody Order order) {
+    public ResponseEntity<Order> updateOrder(@PathVariable("id") long id, @Valid @RequestBody Order order) {
         Optional<Order> orderData = orderRepository.findById(id);
-
         if (orderData.isPresent()) {
 
-            // a finir
             Order _order = orderData.get();
-
+            _order.setName(order.getName());
+            _order.setAddress(order.getAddress());
+            _order.setPostal_code(order.getPostal_code());
+            _order.setCity(order.getCity());
+            _order.setCountry(order.getCountry());
             return new ResponseEntity<>(orderRepository.save(_order), HttpStatus.OK);
         } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            order.setId(id);
+            return new ResponseEntity<>(orderRepository.save(order), HttpStatus.OK);
         }
     }
+
+    /*
+        PATCH NON IMPLEMENTE DEPENDANCE JSON Patch Format
+     */
 
     @DeleteMapping("/orders/{id}")
     public ResponseEntity<HttpStatus> deleteOrder(@PathVariable("id") long id) {
         try {
-            orderRepository.deleteById(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            Optional<Order> orderData = orderRepository.findById(id);
+            if(orderData.isPresent()){
+                orderRepository.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }else{
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
